+++ rastervision_pytorch_learner/rastervision/pytorch_learner/learner.py
@@ -78,7 +78,7 @@ class Learner(ABC):
         Args:
             cfg: configuration
             tmp_dir: root of temp dirs
-            model_def_path: a local path to a directory with a hubcnf.py
+            model_def_path: a local path to a directory with a hubconf.py
             model_path: a local path to model weights. If provided, the model is loaded
                 and it is assumed that this Learner will be used for prediction only.
         """
